requirements.txt
psutil
Pillow
pyautogui
